More fixes for WSL

+++ lib/modules/k2s/k2s.cluster.module/setup/setupcluster.module.psm1
@@ -232,7 +232,7 @@ function Join-WindowsNode {
         $content = (Get-Content -path $joinConfigurationTemplateFilePath -Raw)
         $content.Replace('__CA_CERT__', $caCertFilePath).Replace('__API__', $apiServerEndpoint).Replace('__TOKEN__', $token).Replace('__SHA__', $hash).Replace('__CRI_SOCKET__', 'npipe:////./pipe/containerd-containerd').Replace('__NODE_IP__', $windowsNodeIpAddress) | Set-Content -Path "$joinConfigurationFilePath"
 
-        $joinCommand = '.\' + "kubeadm join $apiServerEndpoint" + ' --node-name ' + $env:COMPUTERNAME + ' --ignore-preflight-errors IsPrivilegedUser,SystemVerification' + " --config `"$joinConfigurationFilePath`""
+        $joinCommand = '.\' + "kubeadm join $apiServerEndpoint" + ' --node-name ' + $env:COMPUTERNAME + ' --ignore-preflight-errors IsPrivilegedUser,SystemVerification' + " --config `"$joinConfigurationFilePath`" -v=5"
 
         Write-Log $joinCommand
 
@@ -243,12 +243,19 @@ function Join-WindowsNode {
         $controlPlaneHostName = Get-ConfigControlPlaneNodeHostname
         $job = Invoke-Expression "Start-Job -ScriptBlock `${Function:Wait-ForNodesReady} -ArgumentList $controlPlaneHostName"
         Set-Location $tempKubeadmDirectory
-        Invoke-Expression $joinCommand 2>&1 | Write-Log
+        $joinOutput = Invoke-Expression $joinCommand 2>&1
+        $joinExitCode = $LASTEXITCODE
+        $joinOutput | Write-Log
+        Write-Log "kubeadm join exit code: $joinExitCode"
         Set-Location ..\..
 
         # print the output of the WaitForJoin.ps1
         Receive-Job $job
         $job | Stop-Job
+        
+        if ($joinExitCode -ne 0) {
+            throw "kubeadm join failed with exit code $joinExitCode"
+        }
 
         # delete path if was created
         Remove-Item -Path $tempKubeadmDirectory\kubeadm.exe
